lagt til mye links

+++ Node2/Node2/temp drives/CAN_bus.h
@@ -12,7 +12,7 @@
 #include <util/delay.h>
 
 struct CAN_message{
-	int id;
+	uint8_t id;
 	uint8_t length;
 	int8_t data[8]; 
 	};
